requirements.txt
pyyaml
websocket_client
python-telegram-bot